requirements-test.txt
# Testing Requirements for ANZX AI Platform
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
requests>=2.31.0
aiohttp>=3.8.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.11.0
python-dotenv>=1.0.0
faker>=19.0.0
factory-boy>=3.3.0
httpx>=0.24.0
asyncpg>=0.28.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
